Move individual config files to the subpackages that make use of them
Instead of having all of them in the main daemon package, move jack.conf
to the -jack subpackage, pipewire-pulse.conf to -pulseaudio subpackage,
and client.conf and client-rt.conf to -libs subpackage.

This makes sure we have the correct config files installed in Fedora
flatpak runtime where the daemon is not installed (but the library
packages are).